idempotence for an exec resource

+++ exercise2.pp
@@ -31,6 +31,7 @@ node "myhostname" {
         command     => "/usr/bin/wget -q 'https://raw.githubusercontent.com/OR1-1/SE1_memory_check/dev001/memory_check.sh' -O /home/monitor/scripts/memory_check",
         creates     => "/home/monitor/scripts/memory_check",
         refreshonly => true,
+        unless => 'test -f /home/monitor/scripts/memory_check',
         notify => File['/home/monitor/scripts/memory_check'],
     }
